Add Login page
Added the Login Page, it compares user credentials to those created manually on the database. If valid login information is presented, then redirects to the Home page.
TO DO: Add functionality to the create an account page that is currently empty. So that users that dont already have an account can create one