Style Login page
Style login page to have a box with half of it having a picture, other half having the login fields. Above the login fields is the logo of our website. The submit button goes from an outlined button to a filled button. The sign in with google button and create an account button change colors when hovered above.